example/llm/llm-icl/requirements1101.txt
absl-py==2.1.0
accelerate==0.21.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anthropic==0.37.1
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.11.2
bitsandbytes==0.39.1
bs4==0.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmake==3.30.5
cohere==3.7.0
contourpy==1.3.0
cycler==0.12.1
datasets==2.16.1
dill==0.3.7
distro==1.9.0
easyinstruct==0.1.2
exceptiongroup==1.2.2
fastavro==1.9.7
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2023.10.0
h11==0.14.0
hanlp==2.1.0b62
hanlp_common==0.0.20
hanlp_downloader==0.0.25
hanlp_trie==0.0.5
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
hydra-core==1.3.2
idna==3.10
importlib_resources==6.4.5
jieba==0.42.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
lexicalrichness==0.5.1
lit==18.1.8
lxml==4.9.2
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
networkx==3.2.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.9.0
packaging==24.1
pandarallel==1.6.5
pandas==2.2.3
parameterized==0.9.0
peft==0.4.0
phrasetree==0.0.9
pillow==11.0.0
propcache==0.2.0
protobuf==3.20.1
psutil==6.1.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==1.10.7
pydantic_core==2.24.2
pynvml==11.5.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rouge-chinese==1.0.3
rouge_score==0.1.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.9.1
sentencepiece==0.1.98
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlitedict==2.1.0
sympy==1.13.1
termcolor==2.5.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.20.1
toposort==1.5
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.45.2
triton==3.1.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
xxhash==3.5.0
yarl==1.15.5
zhconv==1.4.3
zipp==3.20.2






langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.0
openai==1.45.0
pypdf==4.3.1
pytest==8.2.2
nltk==3.9.1
transformers==4.44.0
torch==2.4.0
accelerate==0.27.2
